+++ client/rest-high-level/src/test/java/org/opensearch/client/ReindexIT.java
@@ -112,10 +112,9 @@ public class ReindexIT extends OpenSearchRestHighLevelClientTestCase {
             reindexRequest.setRefresh(true);
             reindexRequest.setRequireAlias(true);
 
-            OpenSearchStatusException exception = expectThrows(
-                OpenSearchStatusException.class,
-                () -> { execute(reindexRequest, highLevelClient()::reindex, highLevelClient()::reindexAsync); }
-            );
+            OpenSearchStatusException exception = expectThrows(OpenSearchStatusException.class, () -> {
+                execute(reindexRequest, highLevelClient()::reindex, highLevelClient()::reindexAsync);
+            });
             assertEquals(RestStatus.NOT_FOUND, exception.status());
             assertEquals(
                 "OpenSearch exception [type=index_not_found_exception, reason=no such index [dest] and [require_alias] request flag is [true] and [dest] is not an alias]",