+++ server/src/main/java/org/opensearch/action/admin/indices/forcemerge/ForceMergeRequest.java
@@ -81,6 +81,8 @@ public class ForceMergeRequest extends BroadcastRequest<ForceMergeRequest> {
      */
     private final String forceMergeUUID;
 
+    private boolean shouldStoreResult;
+
     /**
      * Constructs a merge request over one or more indices.
      *
@@ -160,6 +162,18 @@ public class ForceMergeRequest extends BroadcastRequest<ForceMergeRequest> {
     public ForceMergeRequest flush(boolean flush) {
         this.flush = flush;
         return this;
+    }
+
+    /**
+     * Should this task store its result after it has finished?
+     */
+    public void setShouldStoreResult(boolean shouldStoreResult) {
+        this.shouldStoreResult = shouldStoreResult;
+    }
+
+    @Override
+    public boolean getShouldStoreResult() {
+        return shouldStoreResult;
     }
 
     @Override